+++ src/main/java/com/paymenttransaction/payment_transaction_manager/config/ApplicationConfig.java
@@ -1,9 +1,10 @@
 package com.paymenttransaction.payment_transaction_manager.config;
 
 import com.paymenttransaction.payment_transaction_manager.application.ports.out.TransactionPort;
-import com.paymenttransaction.payment_transaction_manager.application.useCases.CreateTransactionUseCaseImpl;
+import com.paymenttransaction.payment_transaction_manager.domain.useCases.CreateTransactionUseCaseImpl;
 import com.paymenttransaction.payment_transaction_manager.domain.services.TransactionService;
 import com.paymenttransaction.payment_transaction_manager.infrastructure.adapters.TransactionAdapter;
+import com.paymenttransaction.payment_transaction_manager.infrastructure.mappers.TransactionEntityMapper;
 import com.paymenttransaction.payment_transaction_manager.infrastructure.repositories.TransactionRepository;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
@@ -39,8 +40,8 @@ public class ApplicationConfig {
     }
 
     @Bean
-    public TransactionAdapter transactionAdapter(TransactionRepository repository){
-        return new TransactionAdapter(repository);
+    public TransactionAdapter transactionAdapter(TransactionRepository repository, TransactionEntityMapper mapper){
+        return new TransactionAdapter(repository, mapper);
     }
 
 }