+++ src/main/java/com/paymenttransaction/payment_transaction_manager/infrastructure/controllers/TransactionController.java
@@ -2,11 +2,18 @@ package com.paymenttransaction.payment_transaction_manager.infrastructure.contro
 
 import com.paymenttransaction.payment_transaction_manager.application.dtos.transaction.TransactionRequestDTO;
 import com.paymenttransaction.payment_transaction_manager.application.dtos.transaction.TransactionResponseDTO;
-import com.paymenttransaction.payment_transaction_manager.application.ports.in.CreateTransactionUseCase;
-import com.paymenttransaction.payment_transaction_manager.application.useCases.CreateTransactionUseCaseImpl;
 import com.paymenttransaction.payment_transaction_manager.domain.models.Transaction;
 import com.paymenttransaction.payment_transaction_manager.domain.services.TransactionService;
 import com.paymenttransaction.payment_transaction_manager.infrastructure.mappers.TransactionMapper;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,21 +22,36 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/v1/transactions")
+@Tag(name = "Transactions", description = "Endpoints for managing transactions")
 public class TransactionController {
+
+    private static final Logger log = LoggerFactory.getLogger(TransactionController.class);
 
     private final TransactionService transactionService;
 
-    private final TransactionMapper mapper = TransactionMapper.INSTANCE;
+    private final TransactionMapper mapper;
 
-    public TransactionController(TransactionService transactionService){
+    public TransactionController(TransactionService transactionService, TransactionMapper mapper){
         this.transactionService = transactionService;
+        this.mapper = mapper;
     }
 
+    @Operation(summary = "Crea una nueva transacción", description = "Este endpoint permite registrar una nueva transacción")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Transacción creada exitosamente",
+                    content = @Content(schema = @Schema(implementation = TransactionResponseDTO.class))),
+            @ApiResponse(responseCode = "400", description = "Solicitud inválida", content = @Content),
+            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = @Content)
+    })
     @PostMapping
-    public ResponseEntity<TransactionResponseDTO> createTransaction(@RequestBody TransactionRequestDTO requestDTO) {
+    public ResponseEntity<TransactionResponseDTO> createTransaction(
+            @Valid @RequestBody TransactionRequestDTO requestDTO) {
+        log.info("Transaction created with data: {}", requestDTO);
 
         Transaction createdTransaction = transactionService.execute(mapper.toModel(requestDTO));
+        TransactionResponseDTO responseDTO = mapper.toResponseDTO(createdTransaction);
 
-        return ResponseEntity.ok(mapper.toResponseDTO(createdTransaction));
+        log.info("Transaction Created successfully: {}", responseDTO);
+        return ResponseEntity.status(201).body(responseDTO);
     }
 }