+++ src/main/java/com/paymenttransaction/payment_transaction_manager/infrastructure/exceptions/InvalidTransactionException.java
@@ -1,0 +1,7 @@
+package com.paymenttransaction.payment_transaction_manager.infrastructure.exceptions;
+
+public class InvalidTransactionException extends RuntimeException {
+    public InvalidTransactionException(String message) {
+        super(message);
+    }
+}