+++ src/main/java/com/paymenttransaction/payment_transaction_manager/infrastructure/exceptions/TransactionNotFoundException.java
@@ -1,0 +1,8 @@
+package com.paymenttransaction.payment_transaction_manager.infrastructure.exceptions;
+
+
+public class TransactionNotFoundException extends RuntimeException {
+    public TransactionNotFoundException(String message) {
+        super(message);
+    }
+}