+++ target/generated-sources/annotations/com/paymenttransaction/payment_transaction_manager/infrastructure/mappers/TransactionEntityMapperImpl.java
@@ -3,12 +3,14 @@ package com.paymenttransaction.payment_transaction_manager.infrastructure.mapper
 import com.paymenttransaction.payment_transaction_manager.domain.models.Transaction;
 import com.paymenttransaction.payment_transaction_manager.infrastructure.entities.TransactionEntity;
 import javax.annotation.processing.Generated;
+import org.springframework.stereotype.Component;
 
 @Generated(
     value = "org.mapstruct.ap.MappingProcessor",
-    date = "2025-01-23T02:11:39-0300",
-    comments = "version: 1.5.5.Final, compiler: javac, environment: Java 17.0.13 (Amazon.com Inc.)"
+    date = "2025-01-25T17:36:29-0300",
+    comments = "version: 1.5.5.Final, compiler: javac, environment: Java 17.0.14 (Amazon.com Inc.)"
 )
+@Component
 public class TransactionEntityMapperImpl implements TransactionEntityMapper {
 
     @Override