+++ CHANGELOG.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.2] - 2025-01-23
+
+### Changed
+
+- README.md updated.
+- Classes and tests.
+
+### Added
+
+- OpenAPI/Swagger.
+- Dev environment.
+- JWT security.
+- Exception Handler and exceptions.
+- Transactionality on adapter.
+- Integration Tests.
+
+
+
 ## [0.0.1] - 2025-01-23
 
 ### Added